Moved output logic to lib.output
